+++ commitizen/commands/changelog.py
@@ -1,12 +1,14 @@
 import os.path
+import re
 from difflib import SequenceMatcher
 from operator import itemgetter
 from typing import Callable, Dict, List, Optional
 
 from packaging.version import parse
 
-from commitizen import bump, changelog, defaults, factory, git, out, version_types
+from commitizen import changelog, defaults, factory, git, out, version_types
 from commitizen.config import BaseConfig
+from commitizen.defaults import DEFAULT_SETTINGS
 from commitizen.exceptions import (
     DryRunExit,
     NoCommitsFoundError,
@@ -16,6 +18,7 @@ from commitizen.exceptions import (
     NotAllowed,
 )
 from commitizen.git import GitTag, smart_open
+from commitizen.tags import make_tag_pattern, tag_from_version
 
 
 class Changelog:
@@ -55,8 +58,8 @@ class Changelog:
             or defaults.change_type_order
         )
         self.rev_range = args.get("rev_range")
-        self.tag_format = args.get("tag_format") or self.config.settings.get(
-            "tag_format"
+        self.tag_format: str = args.get("tag_format") or self.config.settings.get(
+            "tag_format", DEFAULT_SETTINGS["tag_format"]
         )
         self.merge_prerelease = args.get(
             "merge_prerelease"
@@ -64,6 +67,11 @@ class Changelog:
 
         version_type = self.config.settings.get("version_type")
         self.version_type = version_type and version_types.VERSION_TYPES[version_type]
+
+        tag_regex = args.get("tag_regex") or self.config.settings.get("tag_regex")
+        if not tag_regex:
+            tag_regex = make_tag_pattern(self.tag_format)
+        self.tag_pattern = re.compile(str(tag_regex), re.VERBOSE | re.IGNORECASE)
 
     def _find_incremental_rev(self, latest_version: str, tags: List[GitTag]) -> str:
         """Try to find the 'start_rev'.
@@ -138,7 +146,7 @@ class Changelog:
         # Don't continue if no `file_name` specified.
         assert self.file_name
 
-        tags = git.get_tags()
+        tags = git.get_tags(pattern=self.tag_pattern)
         if not tags:
             tags = []
 
@@ -148,7 +156,7 @@ class Changelog:
             changelog_meta = changelog.get_metadata(self.file_name)
             latest_version = changelog_meta.get("latest_version")
             if latest_version:
-                latest_tag_version: str = bump.normalize_tag(
+                latest_tag_version: str = tag_from_version(
                     latest_version,
                     tag_format=self.tag_format,
                     version_type_cls=self.version_type,